refactor util.py
