+++ src/leopard_em/backend/core_refine_template.py
@@ -17,8 +17,8 @@ from leopard_em.backend.cross_correlation import (
 )
 from leopard_em.backend.distributed import run_multiprocess_jobs
 from leopard_em.backend.utils import normalize_template_projection
-from leopard_em.pydantic_models.utils import calculate_ctf_filter_stack_full_args
 from leopard_em.utils.cross_correlation import handle_correlation_mode
+from leopard_em.utils.ctf_utils import calculate_ctf_filter_stack_full_args
 
 # This is assuming the Euler angles are in the ZYZ intrinsic format
 # AND that the angles are ordered in (phi, theta, psi)
