+++ src/leopard_em/utils/__init__.py
@@ -1,6 +1,7 @@
 """Utilities submodule for various data and pre- and post-processing tasks."""
 
 from .cross_correlation import handle_correlation_mode
+from .ctf_utils import calculate_ctf_filter_stack
 from .data_io import (
     load_mrc_image,
     load_mrc_volume,
@@ -10,9 +11,22 @@ from .data_io import (
     write_mrc_from_numpy,
     write_mrc_from_tensor,
 )
+from .image_processing import (
+    preprocess_image,
+    volume_to_rfft_fourier_slice,
+)
+from .search_utils import (
+    cs_to_pixel_size,
+    get_cs_range,
+    get_search_tensors,
+)
 
 __all__ = [
+    # Cross correlation
     "handle_correlation_mode",
+    # CTF utilities
+    "calculate_ctf_filter_stack",
+    # Data I/O
     "read_mrc_to_numpy",
     "read_mrc_to_tensor",
     "write_mrc_from_numpy",
@@ -20,4 +34,11 @@ __all__ = [
     "load_mrc_image",
     "load_mrc_volume",
     "load_template_tensor",
+    # Image processing
+    "preprocess_image",
+    "volume_to_rfft_fourier_slice",
+    # Search utilities
+    "get_search_tensors",
+    "get_cs_range",
+    "cs_to_pixel_size",
 ]
